+++ cgi/studentDelete.js
@@ -92,6 +92,5 @@ http.createServer(function (req, res) {
     } catch (error) {
         console.log(error); //
     }
-}).listen(8390);
+}).listen(8084);
 console.log('Server started on localhost:8387; press Ctrl-C to terminate...!'); //
-// 99 / 98
