+++ cgi/studentEdit.js
@@ -116,6 +116,5 @@ http.createServer(function (req, res) {
     } catch (error) {
         console.log(error); //
     }
-}).listen(8386);
-console.log('Server running at http://127.0.0.1:8386/'); //
-// 122 / 123
+}).listen(8085);
+console.log('Server running at http://127.0.0.1:8085/'); //
