+++ cgi2/select.js
@@ -21,5 +21,5 @@ http.createServer(function (req, res) {
     } catch (error) {
         console.log(error); //
     }
-}).listen(8388);
-console.log('Server running at http://127.0.0.1:8388/'); //
+}).listen(8087);
+console.log('Server running at http://127.0.0.1:8087/'); //
