+++ cgi2/selectPhone.js
@@ -34,5 +34,5 @@ http.createServer(function (req, res) {
     } catch (error) {
         console.log(error); //
     }
-}).listen(8088);
-console.log('Server running at http://127.0.0.1:8388/'); //
+}).listen(8083);
+console.log('Server running at http://127.0.0.1:8083/'); //
