+++ data/data.json
@@ -1,1 +1,50 @@
-[{"uid":1,"name":"BugzFN","phone":"347-803-9647","secondPhone":"569-283-4656","img":"1.jpg","morePhones":"y"},{"uid":2,"name":"ProofFN","phone":"135-546-5487","secondPhone":"259-135-3645","img":"2.jpg","morePhones":"y"},{"uid":3,"name":"BizarreFN","phone":"237-457-8658","secondPhone":null,"img":"3.jpg","morePhones":"n"},{"uid":4,"name":"EminemFN","phone":"456-324-9346","secondPhone":"267-445-3466","img":"4.jpg","morePhones":"y"},{"uid":5,"name":"KunivaFN","phone":"556-162-4567","secondPhone":"","img":"5.jpg","morePhones":"n"},{"uid":6,"name":"DenaunPorterFN","phone":"564-536-5894","secondPhone":"785-856-3789","img":"6.jpg","morePhones":"y"}]
+[
+    {
+        "uid": 1,
+        "name": "BugzFN",
+        "phone": "347-803-9647",
+        "secondPhone": "569-283-4656",
+        "img": "1.jpg",
+        "morePhones": "y"
+    },
+    {
+        "uid": 2,
+        "name": "ProofFN",
+        "phone": "135-546-5487",
+        "secondPhone": "259-135-3645",
+        "img": "2.jpg",
+        "morePhones": "y"
+    },
+    {
+        "uid": 3,
+        "name": "BizarreFN",
+        "phone": "237-457-8658",
+        "secondPhone": null,
+        "img": "3.jpg",
+        "morePhones": "n"
+    },
+    {
+        "uid": 4,
+        "name": "EminemFN",
+        "phone": "456-324-9346",
+        "secondPhone": "267-445-3466",
+        "img": "4.jpg",
+        "morePhones": "y"
+    },
+    {
+        "uid": 5,
+        "name": "KunivaFN",
+        "phone": "556-162-4567",
+        "secondPhone": "",
+        "img": "5.jpg",
+        "morePhones": "n"
+    },
+    {
+        "uid": 6,
+        "name": "DenaunPorterFN",
+        "phone": "564-536-5894",
+        "secondPhone": "785-856-3789",
+        "img": "6.jpg",
+        "morePhones": "y"
+    }
+]
