+++ data/phone.json
@@ -1,1 +1,31 @@
-[[1,"347-803-9647","569-283-4656","456-234-2457"],[2,"135-546-5487","259-135-3645"],[3,"237-457-8658"],[4,"456-324-9346","267-445-3466"],[5,"556-162-4567"],[6,"564-536-5894","785-856-3789"]]
+[
+    [
+        1,
+        "347-803-9647",
+        "569-283-4656",
+        "456-234-2457"
+    ],
+    [
+        2,
+        "135-546-5487",
+        "259-135-3645"
+    ],
+    [
+        3,
+        "237-457-8658"
+    ],
+    [
+        4,
+        "456-324-9346",
+        "267-445-3466"
+    ],
+    [
+        5,
+        "556-162-4567"
+    ],
+    [
+        6,
+        "564-536-5894",
+        "785-856-3789"
+    ]
+]
