+++ www/error.htm
@@ -1,0 +1,13 @@
+<html>
+
+<head>
+    <style>
+        a {
+            font-size: 40px;
+        }
+    </style>
+    <p> <img src="http://localhost:8082/?img=error.jpg&type=content" width="500" lenght="500">
+        &nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;&nbsp;<a href="http://localhost:8080/?type=index">HOME</a> </p>
+</head>
+
+</html>
