+++ www/index.htm
@@ -57,14 +57,14 @@
     <script>
         // -------------ToDo: FUNCTION FOR: WHEN YOU CLICK ON THE ADD MORE PHONES; (NOT USED) (SCRAPPED);
 
-        // 8385 - studentAdd.js;
-        // 8088 - selectPhone.js;
-        // 8390 - studentDelete.js;
-        // 8386 - studentEdit.js;
-        // 8388 - select.js;
+        // 8086 - studentAdd.js;
+        // 8083 - selectPhone.js;
+        // 8084 - studentDelete.js;
+        // 8085 - studentEdit.js;
+        // 8087 - select.js;
         // 8080 - index.js;
-        // 8091 - indexPhone.js;
-        // 8384 - image.js;
+        // 8081 - indexPhone.js;
+        // 8082 - image.js;
 
         // ---------------------------------------------------------------------------GLOBAL VARIABLES;
 
@@ -87,7 +87,7 @@
         function onStart() {
             document.getElementById("formDiv").style.display = "none";
             document.getElementById("addPhoneDiv").style.display = "none";
-            var url = "http://localhost:8388/?";
+            var url = "http://localhost:8087/?";
             xhttp.open("GET", url, true);
 
             xhttp.onreadystatechange = function () {
@@ -134,10 +134,10 @@
                     } else if (j == 2) {
                         if (mainArray[i].morePhones == "y") {
                             var secondNumber = mainArray[i].secondPhone[0] + mainArray[i].secondPhone[1];
-                            cellText = "<a href=\"http://localhost:8091/?studentUid=" + mainArray[i].uid + "\">" + mainArray[i].phone + "; " + secondNumber + "...</p>";
+                            cellText = "<a href=\"http://localhost:8080/?studentUid=" + mainArray[i].uid + "&type=phone\">" + mainArray[i].phone + "; " + secondNumber + "...</p>";
                             cell.innerHTML = cellText;
                         } else {
-                            cellText = "<a href=\"http://localhost:8091/?studentUid=" + mainArray[i].uid + "\">" + mainArray[i].phone + "; ...</p>";
+                            cellText = "<a href=\"http://localhost:8080/?studentUid=" + mainArray[i].uid + "&type=phone\">" + mainArray[i].phone + "; ...</p>";
                             cell.innerHTML = cellText;
                         }
                     }
@@ -146,11 +146,11 @@
                         image.alt = "Image";
                         image.width = "150";
                         image.height = "100";
-                        image.src = "http://localhost:8384/?img=" + mainArray[i].img + "&type=data";
+                        image.src = "http://localhost:8082/?img=" + mainArray[i].img + "&type=data";
                         cell.appendChild(image);
                     } else if (j == 4) {
-                        cellText = "<p> <img src=\"http://localhost:8384/?img=edit&type=content\" class=\"image\" height=\"30\" size=\"30\" alt=\"edit\" onclick=\"onClickEdit(" + i + ")\"> ";
-                        cellText = cellText + "<img src=\"http://localhost:8384/?img=delete&type=content\" class=\"image\" height=\"30\" size=\"30\" alt=\"del\" onclick=\"onClickDelete(" + i + ")\"></p>";
+                        cellText = "<p> <img src=\"http://localhost:8082/?img=edit-icon.png&type=content\" class=\"image\" height=\"30\" size=\"30\" alt=\"edit\" onclick=\"onClickEdit(" + i + ")\"> ";
+                        cellText = cellText + "<img src=\"http://localhost:8082/?img=delete-icon.png&type=content\" class=\"image\" height=\"30\" size=\"30\" alt=\"del\" onclick=\"onClickDelete(" + i + ")\"></p>";
                         cell.innerHTML = cellText;
                     }
                     row.appendChild(cell);
@@ -183,7 +183,7 @@
 
         function confirmFuncDel(row) {
             if (confirm("Are you sure you want to delete this number?")) {
-                var url = "http://localhost:8390/?";
+                var url = "http://localhost:8084/?";
                 var param = "number=" + mainArray[row].uid;
                 xhttp.open("GET", url + param, true);
                 xhttp.onreadystatechange = function () {
@@ -206,7 +206,7 @@
             newArray.name = document.getElementById("nameInputId").value;
             newArray.phone = document.getElementById("phoneInputId").value;
             hideShowAfterSubmit();
-            var url = "http://localhost:8386/?";
+            var url = "http://localhost:8085/?";
             var param = "edit=" + JSON.stringify(newArray);
             xhttp.open("GET", url + param, true);
             xhttp.onreadystatechange = function () {
@@ -227,7 +227,7 @@
             var newArray = { "uid": -1, "name": document.getElementById("nameInputId").value, "phone": document.getElementById("phoneAddId").value, "secondPhone": "", "img": "1.jpg", "morePhones": "n" };
 
             hideShowAfterSubmit();
-            var url = "http://localhost:8385/?";
+            var url = "http://localhost:8086/?";
             var param = "addNew=" + JSON.stringify(newArray) + "&type=index";
             xhttp.open("GET", url + param, true);
             xhttp.onreadystatechange = function () {
@@ -281,7 +281,7 @@
         <!-- / -->
     </p>
     <div id="imageDiv">
-        <img src="http://localhost:8384/?img=add&type=content" class="image2" height="50" size="50" alt="add"
+        <img src="http://localhost:8082/?img=add-icon.png&type=content" class="image2" height="50" size="50" alt="add"
             onclick="addNew()">
     </div>
     <div id="formDiv">
