+++ www/indexPhone.htm
@@ -62,14 +62,14 @@
     </style>
     <script>
 
-        // 8385 - studentAdd.js;
-        // 8088 - selectPhone.js;
-        // 8390 - studentDelete.js;
-        // 8386 - studentEdit.js;
-        // 8388 - select.js;
+        // 8086 - studentAdd.js;
+        // 8083 - selectPhone.js;
+        // 8084 - studentDelete.js;
+        // 8085 - studentEdit.js;
+        // 8087 - select.js;
         // 8080 - index.js;
-        // 8091 - indexPhone.js;
-        // 8384 - image.js;
+        // 8081 - indexPhone.js;
+        // 8082 - image.js;
 
         // --------------------------------------------------------------------------------VARIABLES;
 
@@ -88,7 +88,7 @@
         function onStart() {
             document.getElementById("formDiv").style.display = "none";
             var numHere = "@uidNumberHere";
-            var url = "http://localhost:8088/?studentUid=" + numHere;
+            var url = "http://localhost:8083/?studentUid=" + numHere;
             addCurrentRow = numHere;
             xhttp.open("GET", url, true);
 
@@ -129,8 +129,8 @@
                         cellText = phoneNumber;
                         cell.innerHTML = cellText;
                     } else if (j == 1) {
-                        cellText = "<img src=\"http://localhost:8384/?img=edit&type=content\" class=\"image\" height=\"30\" size=\"30\" alt=\"edit\" onclick=\"onClickEdit(" + i + ")\">";
-                        cellText = cellText + "<img src=\"http://localhost:8384/?img=delete&type=content\" class=\"image\" height=\"30\" size=\"30\" alt=\"del\" onclick=\"onClickDelete(" + i + ")\">";
+                        cellText = "<img src=\"http://localhost:8082/?img=edit-icon.png&type=content\" class=\"image\" height=\"30\" size=\"30\" alt=\"edit\" onclick=\"onClickEdit(" + i + ")\">";
+                        cellText = cellText + "<img src=\"http://localhost:8082/?img=delete-icon.png&type=content\" class=\"image\" height=\"30\" size=\"30\" alt=\"del\" onclick=\"onClickDelete(" + i + ")\">";
                         cell.innerHTML = cellText;
                     }
                     row.appendChild(cell);
@@ -158,7 +158,7 @@
         // -------------CONFIRM  THAT YOU WANT TO DELETE THE NUMBER;
         function confirmFunc(iValue) {
             if (confirm("Are you sure you want to delete this number?")) {
-                var url = "http://localhost:8390/?";
+                var url = "http://localhost:8084/?";
                 mainArray.splice(iValue, 1);
                 var param = "phoneNumber3523=" + JSON.stringify(mainArray) + "&iValueCurrent=" + iValue;
                 xhttp.open("GET", url + param, true);
@@ -188,7 +188,7 @@
             if (currentFunction == "edit") {
                 hideShowAfterSubmit();
                 mainArray[currentRow] = document.getElementById("phoneText").value;
-                var url = "http://localhost:8386/?";
+                var url = "http://localhost:8085/?";
                 var param = "editFromIndexPhone35252007=" + JSON.stringify(mainArray);
                 xhttp.open("GET", url + param, true);
                 xhttp.onreadystatechange = function () {
@@ -204,7 +204,7 @@
             } else if (currentFunction == "add") {
                 hideShowAfterSubmit();
                 var newArray = document.getElementById("phoneText").value;
-                var url = "http://localhost:8385/?";
+                var url = "http://localhost:8086/?";
                 var param = "addNew=" + JSON.stringify(newArray) + "&type=phone&currentPart=" + addCurrentRow;
                 xhttp.open("GET", url + param, true);
                 xhttp.onreadystatechange = function () {
@@ -237,11 +237,11 @@
 
 <body onload="onStart()">
     <div id="backDiv">
-        <h3><a href="http://localhost:8080/">BACK</a></h3>
+        <h3><a href="http://localhost:8080/?type=index">BACK</a></h3>
     </div>
     <p id="error"></p>
     <div id="imageDiv">
-        <img src="http://localhost:8384/?img=add&type=content" class="image2" height="40" size="40" alt="add"
+        <img src="http://localhost:8082/?img=add-icon.png&type=content" class="image2" height="40" size="40" alt="add"
             onclick="onClickAdd()">
     </div>
     <div id="formDiv">
